Adjust app for API lowercase change
API now returns fields in camelcase while the first letter is lowercase, I think it is good to do this now before it is too late.
Why we did uppercase in the beginning  ? RavenDB(DocumentDB we use) did that in the docs, but we think now it is not good for us.