Add Custom url for share rate (#19)
* Update: fluttertoast version

* Enhance: failures.dart

* Enhance: translation cubit

* Enhance: remove dead code

* Add: toggle language button

* Add: translation

* Handle: tab bar screen

* Add: new translations and update translation screen

This commit introduces additional translations for Arabic and English, including terms like "sun," "justice," and "good morning." The translation screen has been updated to utilize these new strings, enhancing the localization feature of the app.

* Enhance: update text direction handling in custom form field

* Add: new translation strings and update translation logic

This commit introduces additional translation strings for both Arabic and English, enhancing the localization of the app. It also updates the translation logic in the TranslationWidget and TranslationCubit to support the new strings, improving the user experience when switching languages.

* Add: new translation strings and update UI components for localization

* Add: additional translation strings for English and Arabic, and update translation logic in the TranslationScreen

* Update: modify translation strings and enhance suggestion handling

* Refactor: improve code readability and formatting in TranslationWidget

* Translate: add "add to favorites" text to localization files

* Update: emit TranslationInitial state in update function

* Refactor: clear translation text field on language switch

* Update: enhance app URLs and add iOS-specific settings in Info.plist

This commit introduces platform-specific app URLs for sharing and rating, improving user experience across iOS and Android. Additionally, it updates the Info.plist with new keys for frame duration and indirect input events, and modifies build settings in project.pbxproj for better code signing and app identification.

* Update version number in pubspec.yaml to 1.3.12+17

---------

Co-authored-by: Muhammad Hesham <muhammadhisham100@gmail.com>