Add dictionary screen (#22)
* Add Dictionary Feature (#21)

This commit introduces a new Dictionary feature to the app, including a DictionaryCubit for managing state, a DictionaryScreen for displaying category words, and associated data models and use cases. The TabBarScreen has been updated to include the new DictionaryScreen, enhancing the app's functionality and user experience. Additionally, the .metadata file has been updated to reflect platform changes.

* Add Dictionary Feature Enhancements

This commit updates the Dictionary feature by integrating the DictionaryCubit into the main application and modifying the DictionaryScreen to load category words dynamically. It also includes improvements to the dependency injection setup by adding a new use case for fetching dictionary category words. Additionally, code formatting has been enhanced for better readability.

* Enhance Dictionary Feature with Category Selection

This commit updates the DictionaryScreen to support dynamic category selection for loading words. It refactors the screen from a StatelessWidget to a StatefulWidget, allowing users to choose categories from a horizontal list. Additionally, error handling is improved in the DictionaryCubit when fetching category words, ensuring a smoother user experience. The AndroidManifest.xml is also updated to enable back navigation callbacks.

* Refactor AppColors and Enhance DictionaryScreen Initialization

This commit refines the AppColors class by adding new color constants for dictionary categories and improving code formatting. Additionally, it updates the DictionaryScreen to load category words during the initialization phase, ensuring the selected category is ready when the screen is displayed. The layout adjustments enhance the user interface for category selection.

* Add Dictionary Category Management and UI Components

This commit introduces a new data model for dictionary categories, enhancing the dictionary feature with a dedicated `DictionaryCategory` class and a static list of categories. It also adds a `CategorySelectorWidget` for user-friendly category selection and a `DictionaryContentWidget` to display translations based on the selected category. Additionally, the `DictionaryScreen` is updated to manage category selection using `SharedPreferences`, ensuring the selected category persists across sessions. This improves the overall user experience and functionality of the dictionary feature.

* Improve error handling in DictionaryCubit for navigator state access

This commit enhances the DictionaryCubit by adding error handling to ensure that the navigator's current state is not null before accessing its context. If the current state is null, an exception is thrown with a clear message, improving the robustness of the application and preventing potential runtime errors.

* Refactor Dictionary Feature to Remove SharedPreferences Dependency

This commit removes the use of SharedPreferences for storing the selected dictionary category in the DictionaryScreen. The category selection logic is simplified by directly managing the selected category state within the widget. Additionally, the onCategorySelected callback is updated to a synchronous function, improving performance and readability. Unused constants related to SharedPreferences are also removed from the consts.dart file.

* Update version